+++ src/main/java/com/blog/som/domain/member/security/service/CustomUserDetailsService.java
@@ -1,0 +1,29 @@
+package com.blog.som.domain.member.security.service;
+
+import com.blog.som.domain.member.entity.MemberEntity;
+import com.blog.som.domain.member.repository.MemberRepository;
+import com.blog.som.domain.member.security.userdetails.LoginMember;
+import com.blog.som.global.exception.ErrorCode;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+
+@Slf4j
+@RequiredArgsConstructor
+@Service
+public class CustomUserDetailsService implements UserDetailsService {
+
+  private final MemberRepository memberRepository;
+
+  @Override
+  public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+    MemberEntity member = memberRepository.findByEmail(username)
+        .orElseThrow(() -> new UsernameNotFoundException(ErrorCode.MEMBER_NOT_FOUND.getDescription()));
+    log.info("인증 성공[ ID : {} ]", member.getEmail());
+
+    return new LoginMember(member);
+  }
+}
